Simultaneous installation with askgpt plugin
When I tried to solve the problem of not being able to install AskGPT simultaneously with GPT-5, I came up with this suggested change. 

 The sole purpose of these changes is to prevent users from receiving errors when installing AskGPT first and then Assistant.

Should we implement these changes? Or can we find a simpler, more streamlined method? Or is there a need to fix this problem?

It was tested on the Kobo Forma.